requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
asgiref==3.8.1
attrs==23.2.0
bcrypt==4.1.2
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.40.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==5.0.4
django-cors-headers==4.3.1
django-filter==24.3
django-rest-swagger==2.2.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
Flask==3.0.2
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
idna==3.7
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.3
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
Mako==1.3.2
MarkupSafe==2.1.5
multidict==6.0.5
openapi-codec==1.3.2
packaging==24.0
pillow==10.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
simplejson==3.19.2
six==1.16.0
SQLAlchemy==2.0.29
sqlparse==0.4.4
swagger-spec-validator==3.0.3
tqdm==4.66.4
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.30.0
Werkzeug==3.0.1
yarl==1.9.4
